// projects/monad/src/schema/schema.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */
import type { String, Int, BigInt, Float, ID, Bytes, Timestamp, Boolean, Int8 } from '@sentio/sdk/store'
import { Entity, Required, One, Many, Column, ListColumn, AbstractEntity, getStore, UpdateValues } from '@sentio/sdk/store'
import { BigDecimal } from '@sentio/bigdecimal'
import { DatabaseSchema } from '@sentio/sdk'







interface UniswapPoolConstructorInput {
  id: String;
  token0: String;
  token1: String;
  fee: BigInt;
  tickSpacing: BigInt;
  hooks: String;
  amount0: BigInt;
  amount1: BigInt;
}
@Entity("UniswapPool")
export class UniswapPool extends AbstractEntity  {

	@Required
	@Column("String")
	id: String

	@Required
	@Column("String")
	token0: String

	@Required
	@Column("String")
	token1: String

	@Required
	@Column("BigInt")
	fee: BigInt

	@Required
	@Column("BigInt")
	tickSpacing: BigInt

	@Required
	@Column("String")
	hooks: String

	@Required
	@Column("BigInt")
	amount0: BigInt

	@Required
	@Column("BigInt")
	amount1: BigInt
  constructor(data: UniswapPoolConstructorInput) {super()}
  
  
  static update(values: UpdateValues<UniswapPoolConstructorInput>): Promise<void> {
    return getStore().update(UniswapPool, values)
  }
}


const source = `type UniswapPool @entity {
  id: String!
  token0: String!
  token1: String!
  fee: BigInt!
  tickSpacing: BigInt!
  hooks: String!
  amount0: BigInt!
  amount1: BigInt!
}
`
DatabaseSchema.register({
  source,
  entities: {
    "UniswapPool": UniswapPool
  }
})